Re-organized `import`s to leverage the barrel exports
