+++ packages/rangelink-vscode-extension/src/__tests__/destinations/PasteDestinationManager.test.ts
@@ -5,8 +5,7 @@ import * as vscode from 'vscode';
 import type { DestinationAvailabilityService } from '../../destinations/DestinationAvailabilityService';
 import type { PasteDestination } from '../../destinations/PasteDestination';
 import { PasteDestinationManager } from '../../destinations/PasteDestinationManager';
-import { AutoPasteResult } from '../../types/AutoPasteResult';
-import { MessageCode } from '../../types/MessageCode';
+import { AutoPasteResult, MessageCode } from '../../types';
 import * as formatMessageModule from '../../utils/formatMessage';
 import {
   configureEmptyTabGroups,
