+++ packages/rangelink-vscode-extension/src/types/index.ts
@@ -1,8 +1,11 @@
 export { ActiveSelections } from './ActiveSelections';
+export { AutoPasteResult } from './AutoPasteResult';
 export type { AvailableDestination } from './AvailableDestination';
 export { BehaviourAfterPaste } from './BehaviourAfterPaste';
+export { MessageCode } from './MessageCode';
+export { PasteContentType } from './PasteContentType';
+export { QuickPickBindResult } from './QuickPickBindResult';
 export type { RangeLinkClickArgs } from './RangeLinkClickArgs';
 export type { RangeLinkTerminalLink } from './RangeLinkTerminalLink';
 export type { SendTextToTerminalOptions } from './SendTextToTerminalOptions';
-export { QuickPickBindResult } from './QuickPickBindResult';
 export { TerminalFocusType } from './TerminalFocusType';
